pyproject.toml
[tool.poetry]
name = "lmwa"
version = "0.1.0"
description = ""
authors = ["Diptanil Saha <diptanil.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask-sqlalchemy = "^3.1.1"
flask-wtf = "^1.2.1"
python-dotenv = "^1.0.1"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
